Heavy updates to README and commands.py. The program is bare- bones but the general form of it is showing up.
README has a bit more information about the project in general.

commands.py has been updated with extensive documentation.  Doc
strings have been added to all functions and to the module itself.
Most of the functions are still stubs. commands.help is fully
functional, and help for any of the commands can be seen.

driver.py causes a basic read-eval-print loop to start. The user
can enter commands to manipulate the list.